ame16 module
Introduced ame16 component, with the public table (mass16.txt)
converted to a 3-column [Z,A,mass] file for quicker and cleaner loading.

Currently the implementation in HZ90 that “loads ame16 data if it
exists, else reverts to mb77 calculation” is sloppy. Loads the table on
every single call to get_mass because the ame16 data isnt global/clever
yet, very inefficient.