ame16 class
ame16 mass table is now integrated as a class, which can then be
created at the top of hz90 or wherever and invoked globally, no data to
constantly track and toss around. Added new functions to hz90 to “try
table for mass and if it’s not there then use mb77 as backup”. Same for
V_N.

Also an aside: re-defined fit parameters at the top of BBP71, since the
ones in constants_def are from MB77 — similar names but very different
values.